feat : 고객 보험금 납부 구현

+++ src/main/java/fourservings_fiveservings/insurance_system_be/team/contract/repository/ContractRepository.java
@@ -2,10 +2,11 @@ package fourservings_fiveservings.insurance_system_be.team.contract.repository;
 
 import fourservings_fiveservings.insurance_system_be.team.contract.entity.common.ApproveStatus;
 import fourservings_fiveservings.insurance_system_be.team.contract.entity.common.Contract;
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContractRepository extends JpaRepository<Contract, Integer> {
+import java.util.List;
+
+public interface ContractRepository extends JpaRepository<Contract, Long> {
 
     List<Contract> findByApproveStatus(ApproveStatus approveStatus);
 }
